requirements.txt
Django==3.2.0
djangorestframework==3.14.0
django-environ==0.10.0
mysqlclient==2.2.0
isort
black
